Refactor UI modules (#7956)
* base_ui refactoring 1
